% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_method.R
\name{combat_type}
\alias{combat_type}
\title{Create a ComBat type tag for S3 dispatch}
\usage{
combat_type(type)
}
\arguments{
\item{type}{Character scalar, typically \code{"univariate"} or \code{"multivariate"}.}
}
\value{
An object whose class is set to \code{type}; primarily used for S3 dispatch.
}
\description{
Convenience constructor for tagging an object with a class used to choose
univariate vs. multivariate ComBat methods.
}
\examples{
class(combat_type("multivariate"))

}
