% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{eigen_error}
\alias{eigen_error}
\title{Mean absolute eigenvalue error}
\usage{
eigen_error(Sigma, Sigma_hat)
}
\arguments{
\item{Sigma}{True covariance matrix.}

\item{Sigma_hat}{Estimated covariance matrix.}
}
\value{
Numeric scalar: mean absolute difference between sorted eigenvalues.
}
\description{
Mean absolute eigenvalue error
}
\examples{
# eigen_error(S, S_hat)
}
\seealso{
Other covariance-metrics: 
\code{\link{evaluate_cov_recovery}()},
\code{\link{frobenius_norm}()},
\code{\link{kl_divergence}()},
\code{\link{mse_cov}()},
\code{\link{spectral_norm}()}
}
\concept{covariance-metrics}
