% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{frobenius_norm}
\alias{frobenius_norm}
\title{Frobenius norm error between covariance matrices}
\usage{
frobenius_norm(Sigma, Sigma_hat)
}
\arguments{
\item{Sigma}{True covariance matrix.}

\item{Sigma_hat}{Estimated covariance matrix.}
}
\value{
Numeric scalar: \eqn{\| \Sigma - \hat\Sigma \|_F}.
}
\description{
Frobenius norm error between covariance matrices
}
\examples{
# frobenius_norm(S, S_hat)
}
\seealso{
Other covariance-metrics: 
\code{\link{eigen_error}()},
\code{\link{evaluate_cov_recovery}()},
\code{\link{kl_divergence}()},
\code{\link{mse_cov}()},
\code{\link{spectral_norm}()}
}
\concept{covariance-metrics}
