% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mse_cov}
\alias{mse_cov}
\title{Mean squared entrywise error (covariance matrices)}
\usage{
mse_cov(Sigma, Sigma_hat)
}
\arguments{
\item{Sigma}{True covariance matrix.}

\item{Sigma_hat}{Estimated covariance matrix.}
}
\value{
Numeric scalar: mean of squared entrywise errors.
}
\description{
Mean squared entrywise error (covariance matrices)
}
\examples{
# mse_cov(S, S_hat)
}
\seealso{
Other covariance-metrics: 
\code{\link{eigen_error}()},
\code{\link{evaluate_cov_recovery}()},
\code{\link{frobenius_norm}()},
\code{\link{kl_divergence}()},
\code{\link{spectral_norm}()}
}
\concept{covariance-metrics}
