% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{spectral_norm}
\alias{spectral_norm}
\title{Spectral norm (operator 2-norm) of error}
\usage{
spectral_norm(Sigma, Sigma_hat)
}
\arguments{
\item{Sigma}{True covariance matrix.}

\item{Sigma_hat}{Estimated covariance matrix.}
}
\value{
Numeric scalar.
}
\description{
Spectral norm (operator 2-norm) of error
}
\details{
For symmetric matrices, equals the largest absolute eigenvalue
of \code{Sigma - Sigma_hat}.
}
\examples{
# spectral_norm(S, S_hat)
}
\seealso{
Other covariance-metrics: 
\code{\link{eigen_error}()},
\code{\link{evaluate_cov_recovery}()},
\code{\link{frobenius_norm}()},
\code{\link{kl_divergence}()},
\code{\link{mse_cov}()}
}
\concept{covariance-metrics}
