% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_method.R
\name{stan_data_prep}
\alias{stan_data_prep}
\title{Build Stan data (S3 generic)}
\usage{
stan_data_prep(type, ...)
}
\arguments{
\item{type}{S3 tag (e.g., \code{combat_type("univariate")} or \code{"multivariate"}).}

\item{...}{Passed to methods.}
}
\description{
Build Stan data (S3 generic)
}
